+++ IdeIntegration/Vs2010Integration/LanguageService/GherkinProcessingScheduler.cs
@@ -10,57 +10,6 @@ using TechTalk.SpecFlow.Vs2010Integration.Utils;
 
 namespace TechTalk.SpecFlow.Vs2010Integration.LanguageService
 {
-    public interface IGherkinProcessingTask
-    {
-        void Apply();
-        IGherkinProcessingTask Merge(IGherkinProcessingTask other);
-    }
-
-    public class DelegateTask : IGherkinProcessingTask
-    {
-        private readonly Action task;
-        private readonly IIdeTracer tracer;
-
-        public DelegateTask(Action task, IIdeTracer tracer)
-        {
-            this.task = task;
-            this.tracer = tracer;
-        }
-
-        public void Apply()
-        {
-            try
-            {
-                task();
-            }
-            catch(Exception exception)
-            {
-                tracer.Trace("Exception: " + exception, "DelegateTask");
-            }
-        }
-
-        public IGherkinProcessingTask Merge(IGherkinProcessingTask other)
-        {
-            return null;
-        }
-    }
-
-    internal class PingTask : IGherkinProcessingTask
-    {
-        public static readonly IGherkinProcessingTask Instance = new PingTask();
-
-        public void Apply()
-        {
-            //nop;
-        }
-
-        public IGherkinProcessingTask Merge(IGherkinProcessingTask other)
-        {
-            return other;
-        }
-    }
-
-
     public class GherkinProcessingScheduler : IDisposable
     {
         private readonly IVisualStudioTracer visualStudioTracer;
@@ -126,8 +75,10 @@ namespace TechTalk.SpecFlow.Vs2010Integration.LanguageService
             }
         }
 
-        private void ApplyTask(IEnumerable<IGherkinProcessingTask> tasks)
+        private void ApplyTask(IGherkinProcessingTask[] tasks)
         {
+            visualStudioTracer.Trace("Applying task batch of {0} tasks", this, tasks.Length);
+
             IGherkinProcessingTask currentTask = null;
             foreach (var task in tasks)
             {
